requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudinary==1.43.0
colorama==0.4.6
cryptography==44.0.2
Django==5.2
django-allauth==65.6.0
django-autoslug==1.9.9
django-cities-light==3.10.1
django-colorfield==0.12.0
django-countries==7.6.1
django-debug-toolbar==5.1.0
django-extensions==3.2.3
django-phonenumber-field==8.0.0
django-taggit==6.1.0
factory_boy==3.3.3
Faker==37.1.0
idna==3.10
iniconfig==2.1.0
MarkupSafe==3.0.2
oauthlib==3.2.2
packaging==24.2
phonenumbers==9.0.2
pillow==11.1.0
pip==25.0.1
pluggy==1.5.0
progressbar2==4.5.0
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.5
pytest-django==4.11.1
python-decouple==3.8
python-utils==3.9.1
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.3
stripe==12.0.0
typing_extensions==4.13.0
tzdata==2025.2
Unidecode==1.3.8
urllib3==2.3.0
Werkzeug==3.1.3
